reports-service/backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlalchemy==1.4.54
hvac
python-jose==3.3.0
psycopg2-binary==2.9.6
